+++ cmd/hidden_lake/adapters/common/consumer/internal/adapted/consumer.go
@@ -10,8 +10,8 @@ import (
 	"time"
 
 	"github.com/number571/go-peer/cmd/hidden_lake/adapters"
+	"github.com/number571/go-peer/pkg/database"
 	net_message "github.com/number571/go-peer/pkg/network/message"
-	"github.com/number571/go-peer/pkg/storage/database"
 	"github.com/syndtr/goleveldb/leveldb"
 )