+++ pkg/database/settings.go
@@ -1,4 +1,4 @@
-package storage
+package database
 
 var (
 	_ ISettings = &sSettings{}